急停和超声波未调通

+++ STM32F103_Ultrasonic/Board/ultrasonic/ultrasonic.h
@@ -54,7 +54,7 @@
 
 typedef struct
 {
-  float distance;
+  int distance;
   uint8_t flag; //上升沿触发标志位
   uint32_t start_time;
   uint32_t end_time;
